requirements.txt
fastapi[standard]
pyjwt
build
twine
python-dotenv